[Fix #9235] Allow `--only` and `--except` to be able to properly qualify cops added by require.
Previously, cop names passed to `--only` or `--except` were qualified immediately, but because the configuration is not yet loaded when `Options.parse` is called, it could not properly resolve cops that are added by an extension or a local `require`. This was generally not a problem because unresolved cops were accepted, but in the case of a cop that shares a name with a default cop, this resulted in raising a warning that the department name was wrong, and in fact loading the wrong cop.

Instead, qualifying cop names is now done lazily, so that they are only qualified right before they are needed. This ensures that the full registry is present and external cops can be properly detected.