Fix `Lint/AmbiguousRegexpLiteral` when regexp is in `=~` expression
The `find_offense_node` method walks up the AST from the regexp node to
find the method call to correct by making sure it has parentheses, but it
doesn't consider that the parent of the regexp node might be an `=~`
operator expression. This doesn't meet the conditions in the method for
continuing the search, so the method returns the `match_with_lvasgn` node
that represents the `=~` expression. As this isn't the kind of node that
the corrector method `add_parentheses` expects it doesn't have an
`arguments` method which `add_parentheses` will call. The solution is to
change the conditions in `find_offense_node` to continue the search if
the node we're looking at is an `match_with_lvasgn` node so we can find a
real `send_type` parent that we can call `arguments` on.