[Fix #8938] Fix `--auto-gen-config` for `Style/AccessModifierDeclarations`
The expected use of the API provided by `ConfigurableEnforcedStyle` is
for the cop to call `correct_style_detected` if the cop detects no
offense or to call one of the other `xxx_style_detected` api methods if
it does detect an offense, and it should also call `add_offense` to
register the offense. Doing this updates `config_to_allow_offenses` which
is used during `--auto-gen-config` to decide how to configure the cop for
the current code under review so that regardless of the styles used the
code passes review.

Prior to this commit, this cop called `correct_style_detected` and
`opposite_style_detected` correctly, but it did not call `add_offense` to
register the offenses correctly. This has two outcomes:

1. During a normal run the cop would only report 1 offense for the whole
   code base under review. When this is fixed it would then report the
   next offense.
2. During a `--auto-gen-config` run the cop would only output
   the first file it encountered an offense in an `Exclude` list which
   means the configuration generated for this cop in `.rubocop_todo.yml`
   will not produce a passing review.

This commit stops treating `opposite_style_detected` as a predicate to
decide if we should call `add_offense` and instead always calls both
methods.