[Fix #8938] Fix `--auto-gen-config` for `Style/MethodDefParentheses`
The expected use of the API provided by `ConfigurableEnforcedStyle` is
for the cop to call `correct_style_detected` if the cop detects no
offense or to call one of the other `xxx_style_detected` api methods if
it does detect an offense, and it should also call `add_offense` to
register the offense. Doing this updates `config_to_allow_offenses` which
is used during `--auto-gen-config` to decide how to configure the cop for
the current code under review so that regardless of the styles used the
code passes review.

Prior to this commit, this cop _only_ called `correct_style_detected`
which means it would always output `EnforcedStyle: <some value>` during
`--auto-gen-config` runs. If the code under review actually has a mix of
styles rather than a single consistent style then the generated config
for this cop in `.rubocop_todo.yml` will not produce a passing review as
instead of an exclusion list, or turning the cop off, it will only have
the `EnforcedStyle: <some value>` config.

This commit calls `unexpected_style_detected`, passing in the relevant
detected incorrect style whenever we call `add_offense` to avoid this
problem. We can't use `opposite_style_detected` because this cop has more
than 2 supported styles.