Make `Lint/NonDeterministicRequireOrder` not to register offense for Ruby 3
`Dir.glob` and `Dir[]` sort the results by default in Ruby 3.0.

- https://bugs.ruby-lang.org/issues/8709
- https://github.com/ruby/ruby/commit/2f1081a451f21ca017cc9fdc585883e5c6ebf618

This PR will make not to register an offense when using Ruby 3.0.

So users don't need this cop when using Ruby 3.0.
It can be deprecated when supporting only Ruby 3.0 and higher.

I will open a PR about a new cop against `Dir.glob.sort` and `Dir[]`.sort,
which became redundant in Ruby 3.0.