[Fix #9290] Fix a false positive for `Layout/SpaceBeforeBrackets`
Fixes #9290.

This PR fixes a false positive for `Layout/SpaceBeforeBrackets`
when using array literal argument for method call.

If receiver is a method call, it is difficult to prevent false positives.
Therefore, This PR change behaviour to detect only when receiver is a variable.

As a result, safe detection is possible.