Fix `StyleGuideBaseURL` not functioning with nested departments
Currently you can set the StyleGuideBaseURL config for cops that ship
within RuboCop or other RuboCop plugins that use a DEPT/COP format. If
the project has nested departments this doesn't work as the message
annotator does not properly split the cop name to determine the
department to lookup.

With this change just the cop name is split from the full DEPT/COP
string allowing the StyleGuideBaseURL to function with nested
departments.

We have nested departments in Cookstyle like `Chef/Deprecations/Foo` and
we'd like to be able to enable this config by default so we can point
our users at our documentation.

Signed-off-by: Tim Smith <tsmith@chef.io>