[#9526] Add `AllowSplatArgument` option to `Style/HashConversion`
Resolves #9526.

This PR add `AllowSplatArgument` option to `Style/HashConversion` and
the option is true by default.
This is because `Hash[*ary]` is not allowed because it is strict.

As discussed in https://github.com/rubocop-hq/rubocop/pull/9478#discussion_r567259779,
the correction code from `Hash[*arg]` is not simply determined.
So, I think it's better to allow it by default as a simple notation.

On the other hand, there may be case where some users want to use the existing
behavior, so this PR have made it optional.