[Fix #9691] Add configuration option InspectBlocks to RedundantLineBreak
Set the default value for this new configuration option to `false`, since we
suspect that many users might consider the `do`..`end` variant with line
breaks more readable.

For RuboCop's own sources we set it to `true`, i.e., the same behavior as
before the parameters existed, at least for now.