Change RedundantLineBreak: InspectBlocks to false internally
Change the setting for internal inspections so that the
Layout/RedundantLineBreak cop does not report multiline blocks. The opinion of
the core team seems to lean towards allowing `do ... end` blocks when the
author feels it's more readable than a single line block with braces.

This commit does not revert any previous changes. That should be done later on
case-by-case basis if desired.