[Fix #5576] Treat relativity of Include parameters correctly
For configuration files whose name begins with `.rubocop`, the paths in
`Include` parameters are relative to the directory of the configuration file.

When another configuration file inherits from such a file, we cannot just copy
the `Include` paths. We must modify them so they select the files that were
originally intended.