[WIP] Add PreferHashRocketsForQuotedSymbols to Style/HashSyntax
Given the elegant syntax for symbol hash keys

    {
      foo: 123,
    }

developers often try to use it in scenarios where they need a string key

    {
      'foo': 123,
    }

not realizing the key remains a symbol, and that they needed to use

    {
      'foo' => 123,
    }

instead.

While there are legitimate uses for quoting a symbol (if it includes
special characters), in most cases this is not the intended use case.

Therefore, the PreferHashRocketsForQuotedSymbols option is added to
Style/HashSyntax, to give the option to disallow the error-prone syntax.