[Fix #9731] Fix two autocorrection issues for `Style/NegatedIfElseCondition`.
Previously, when swapping the `if` and `else` branches, if one of the branches only had a single statement, the branch would end up not being spaced out properly.

Additionally, when there is a node with a comment inside the `if` that has an identical node with a comment *before* the `if` statement, `CommentsHelp.begin_pos_with_comment` returns an incorrect range, which was causing a clobbering error. I fixed this by creating ranges of the entire body of the if and else branches, so that they can just be swapped without relying on `ProcessedSource#ast_with_comments`.