[Fix #7592] Add cop Layout/LineEndStringConcatenationIndentation
The cop Layout/MultilineOperationIndentation doesn't view the
backslash as an operator, which is correct. Therefore, a new cop
is needed to inspect character literals broken up into multiple
lines with backslashes at the ends of the lines.

The new cop deals exclusively with string literals, not any other
constructs where backslashes might be used to "glue" lines
together.