Add new `Lint/EmptyInPattern` cop
This PR adds new `Lint/EmptyInPattern` cop for Ruby 2.7's pattern matching.
It checks for the presence of `in` branches without a body.

```ruby
# bad
case condition
in [a]
  do_something
in [a, b]
end

# good
case condition
in [a]
  do_something
in [a, b]
  nil
end

# good - AllowComments: true (default)
case condition
in [a]
  do_something
in [a, b]
  # noop
end

# bad - AllowComments: false
case condition
in [a]
  do_something
in [a, b]
  # noop
end
```

This cop is similar to `Lint/EmptyWhen`, but with different supported
syntax and Ruby version (requires 2.7 or higher).

And this PR use https://github.com/rubocop/rubocop-ast/pull/183 feature,
so it requires RuboCop AST 1.6.0 or higher.