Support Ruby 2.7's pattern matching syntax for `Layout/SpaceAroundKeyword`
This PR supports Ruby 2.7's pattern matching syntax for `Layout/SpaceAroundKeyword`.

The following is a technical note.

There are the following differences between `case ... when` and `case ... in`.
So there is a TODO test for this patch's test and I confirmed this syntax below:
https://bugs.ruby-lang.org/issues/17925

An answer will determine whether to enable or discard the commented out test in the future.