Support pattern matching for `Style/IdenticalConditionalBranches` cop
This PR supports Ruby 2.7's pattern matching syntax for `Style/IdenticalConditionalBranches` cop.

This `Style/IdenticalConditionalBranches` cop is detecting for `if` and` case`.
Therefore, I think about detecting `case-match` with the cop instead of a new cop.

It may also be suitable for the cop's role of detecting duplicated conditional branches.