Clarify that you can raise exceptions without arguments
Whilst `MyCustomError.new` violates this Rubocop rule, the documentation implies the only way of fixing it is to pass an argument representing the exception message, whereas actually if your custom exception has a default message, it's quite possible you just want to use the default message. This can be achieved without violating this Rubocop rule, by omitting the call to `.new`.