[Fix #8469] Add cop Layout/SpaceAroundSingletonClassOperator
I chose the name SpaceAroundSingletonClassOperator rather than
SpaceAroundSingletonInheritence, which was suggested in the feature
request. Although I'm not sure that << is actually an operator, I am pretty
sure that inheritance is not involved. The operator (for lack of a better term)
opens up the singleton class, or eigenclass, so methods can be added to it.