[Fix #9752] Improve error message for top level department
When there are cop departments that seem to have a structure, like in
the cookstyle gem where departments are named `Chef/Style`,
`Chef/Correctness`, etc., it's easy to assume that there is a `Chef`
department that can disable all departments under it. This is not the
case, but the error message
`Error: unrecognized cop Chef found in .cookstyle.yml` is not
informative enough.

If an unrecognized cop or department is the first part of existing
departments, we list the existing departments. This should help the user
realize what needs to be changed.

Also, we change the wording of `unrecognized cop` to
`unrecognized cop or department`, and extend the did-you-mean
functionality to also look for similar department names.