[Fix #9946] Avoid slow regexp match in Style/CommentedKeyword
The regular expression we use for finding the keyword on the commented
line is very slow if the comment is long. We need to reverse the order
of the two conditions so we only do the regexp matching after we've
discovered that the line begins with one of the keywords we want to
check.