Make `Layout/LeadingCommentSpace` aware of `#:nodoc`
This PR makes `Layout/LeadingCommentSpace` aware of `#:nodoc`.

The following mention is the trigger.
https://twitter.com/kamipo/status/1421031611588485121

It seems that there was no definitive comment on whether to allow `#:nodoc:`.

- https://github.com/rubocop/rubocop/issues/277
- https://github.com/rubocop/rubocop/commit/db79146

I've look RDoc doc's example of Ruby 3.0 and it looks to include leading comment space.

```ruby
module MyModule # :nodoc:
  class Input
  end
end

module OtherModule # :nodoc: all
  class Output
  end
end
```

https://docs.ruby-lang.org/en/3.0.0/RDoc/Markup.html#class-RDoc::Markup-label-Controlling+what+is+documented

I also considered making it an option for the presence or absence of
leading comment space in `#:nodoc:`, but I think it would be better to
consistent it that space is included when possible as the role of cop.