Try fix CI errors on Windows
Not sure why the failure was specific to Windows, but I think the reason
is that these specs were parsing an invalid lockfile. Since the `specs:`
section didn't exist, bundler was failing to parse the lockfile source,
and then trying to use it while parsing specs, resulting in a method
being called on `nil`.