Docs: Interpreting ABCSize
When I introduce ABC to people, they don't know how to interpret the numbers. Having a scale like this might help.

I'm basing this proposed scale partly on my own experience, and partly on CodeClimate, which uses a similar metric called `flog`.

> What’s a good Flog score? It’s subjective, of course, but Jake Scruggs .. suggested that scores above 20 indicate the method may need refactoring, and above 60 is dangerous. Similarly, Code Climate will flag methods with scores above 25, and considers anything above 60 “very complex”.

We can't directly compare `flog` scores and ABC, but the above is still a useful reference in determining our own scale.