Add spell checking GitHub Actions workflow
This PR adds spell checking GitHub Actions workflow and fixes typos.

The following two tools are used for spell checking.

- misspell (requires Golang) ... https://github.com/client9/misspell
- codespell (requires Python) ... https://github.com/codespell-project/codespell

For contributors who don't have Golang and Python installed, it is only
CONTRIBUTING.md and not run with `rake` task.

This GitHub Actions configuration is based on the Rails's spell checking workflow.
https://github.com/rails/rails/blob/v7.0.0.alpha2/.github/workflows/lint.yml