[Fix #10147] Fix `Lint/ElseLayout` for else nodes with a single line body
The ElseLayout cop, in spirit, is trying to prevent accidentally using
else when a user meant to use elsif based on the fact that they used
else in a similar way as elsif (by passing an argument to it).

However; the syntax of having a compact single line else statement with
no other body is valid and likely shouldn't be flagged as problematic.

This commit stops ElseLayout cop from reporting an offense when all the
else branch's children are on the same line as the else node itself,
while still reporting errors if there are children on other lines in the
else branch (which is likely a mistake)