[Fix #10241] Fix `last_column` value for `JSONFormatter`
Fixes #10241.

This PR fixes `last_column` value for `JSONFormatter`.

The minimum value of `start_column: real_column` is 1. So, the minimum value of `last_column` should be 1.
And non-zero value of `last_column` should be used as is.

This patch to `last_column` seems to be required only for `JSONFormatter` and not required for other formatters.

It solves a potential `JSONFormatter` issue for cops that use `add_global_offense`.