[Fix #10339] Support auto-correction for `Naming/BlockForwarding`
Fixes #10339 and Follow up to https://github.com/rubocop/rubocop/pull/10290#issuecomment-999708816.

This PR supports auto-correction for `EnforcedStyle: explicit` of
`Naming/BlockForwarding`.

User can specify the block variable name for auto-correction with `BlockForwardingName`.
The default variable name is `block`. If the name is already in use, it
will not be auto-corrected.