Fix a false positive for `Style/CollectionCompact`
This PR fixes a false positive for `Style/CollectionCompact` when
without receiver for bad methods.

If no receiver is specified, it may be used in a custom method or
open class. For custom methods, there is no guarantee that `compact`
corresponding to `reject` will exist.

The following is a use case of false positive.
https://github.com/artichoke/artichoke/pull/1582/commits/250bf7e

In fact, without a receiver may lead to false positives for these cases.