[Fix #8032] Improve argument alignment correction for kwargs
This fixes detection of incorrect indentation for the case where
- ArgumentAlignment is configured with EnforcedStyle
  'with_fixed_indentation'
- A method call has both regular arguments and keyword arguments
- The first keyword argument is on the same line as at least one of the
  regular arguments

In the original code, the keyword arguments would be seen as one single
argument, and because it starts on the same line as the previous
argument, alignment adjustment would be skipped.

Also, because of the fixed indentation style, HashAlignment always skips
adjustment of the keyword argument block. This means neither cop would
correct the alignment.

This change adjusts this behaviour by treating each keyword argument as
a separate argument if EnforcedStyle is 'with_fixed_indentation'.

This has an additional effect in that each keyword argument will be
separately annotated.