[Fix #10311] Expand send nodes correctly
When Layout/RedundantLineBreak inspects a send node, it starts by
expanding the node to inspect. It does this by traversing the AST
upwards as long as the parent is a send node of a block node.

The problem was that if the inspected node is a single node inside
a do..end block, we shouldn't move to the parent block and inspect
it for redundant line breaks. It's only send nodes before the
do..end that we want to expand in order to inspect "the whole
expression". So we should check what kind of block/send
relationship it is before moving upwards in the AST.