[Fix #10468] Fix a false positive for `Style/FileWrite`
Fixes #10468

This PR fixes a false positive for `Style/FileWrite`
when a splat argument is passed to `f.write`.

There are different method signatures between `File.write` (class method)
and `File#write` (instance method).

- https://ruby-doc.org/core-3.1.1/IO.html#method-c-write
- https://ruby-doc.org/core-3.1.1/IO.html#method-i-write

This PR accepts the case of using splat argument as an argument for `File#write`
because static analysis does not know the contents of the splat argument.