[Fix #10464] Fix an incorrect autocorrect for `Lint/IncompatibleIoSelectWithFiberScheduler`
Fixes #10464.

This PR fixes an incorrect autocorrect for `Lint/IncompatibleIoSelectWithFiberScheduler`
when using `IO.select` with read (or write) argument and using return value.

When the method is successful the return value of `IO.select` is `[[IO]]`,
and the return value of `io.wait_readable` and `io.wait_writable` are `self`.

They are not auto-corrected when assigning a return value because these types are different.
It's up to user how to handle the return value.