[Fix #10310] Add InspectBlocksOnAllLevels config param
Add a configuration parameter to Layout/RedundantLineBreak to
signal whether blocks on any level (depth) should be inspected.
If false, only the innermost level of blocks within blocks will
be inspected. If true, all levels will be inspected.

To achieve backwards compatibility, the default value for the
new parameter is set to true.