[FIx #10356] Add `AllowConsecutiveConditionals` option to `Style/GuardClause`
This PR add `AllowConsecutiveConditionals` option to
`Style/GuardClause` and the option is false by default.

I'd prefer to keep both if statements
if triggered offense at the same depth.

This option works only if the preceding expression
is a conditional at the same depth.

Since the existing behavior isn't erroneous,
this PR have made it optional and set default to false.