Fix a false positive for `Style/DoubleNegation`
This PR fixes the following false positive for `Style/DoubleNegation`
when inside returned conditional clauses with Ruby 2.7's pattern matching.

The following is a reproduction case.

```ruby
def foo?
  case condition
  in foo
    !!foo
  in bar
    !!bar
  else
    !!baz
  end
end
```

It is expected that an offense will not be registered, but in actually
an offense will be registered.

```console
% bundle exec rubocop --only Style/DoubleNegation
(snip)

Inspecting 2 files
C.

Offenses:

example.rb:4:5: C: [Correctable] Style/DoubleNegation: Avoid the use of double negation (!!).
    !!foo
    ^
example.rb:6:5: C: [Correctable] Style/DoubleNegation: Avoid the use of double negation (!!).
    !!bar
    ^

2 files inspected, 2 offenses detected, 2 offenses auto-correctable
```

This is a follow-up to the same issue as #10474 and requires RuboCop AST 1.7 to fix.
Therefore it bumps RuboCop AST required to use https://github.com/rubocop/rubocop-ast/pull/227
to 1.7 or higher.